[webview][multi-profile] Throw IllegalArgumentException on Attempt to Delete Default Profile
Bug: 1446913
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4839354
Commit-Queue: Wayne Jackson Jr. <wbjacksonjr@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1193573}

+++ android_webview/java/src/org/chromium/android_webview/AwBrowserContext.java
@@ -233,10 +233,15 @@ public class AwBrowserContext implements BrowserContextHandle {
      * <p>
      * Name must be non-null and valid Unicode.
      *
-     * @throws IllegalStateException if trying to delete the default profile or a profile which is
-     *                               in use.
+     * @throws IllegalArgumentException if trying to delete the default profile.
+     * @throws IllegalStateException if trying to delete a profile which is in use.
      */
-    public static boolean deleteNamedContext(String name) {
+    public static boolean deleteNamedContext(String name)
+            throws IllegalArgumentException, IllegalStateException {
+        final String defaultContextName = AwBrowserContextJni.get().getDefaultContextName();
+        if (name.equals(defaultContextName)) {
+            throw new IllegalArgumentException("Cannot delete the default profile");
+        }
         return AwBrowserContextJni.get().deleteNamedContext(name);
     }
